lambdas/airport-lookup/requirements.txt
glide-py==1.0.0
